add detect_watering

+++ wapabawama_ros/src/system/include/system.hpp
@@ -53,7 +53,7 @@ class System {
     int getMotorPos(int);
     double getGantrySpd(int);
     //valve function
-    int setFlow(int ,float);
+    int setFlow(int ,double);
     int setPWM(int ,int);
     int returnState(int);
     int closeValve();
@@ -74,6 +74,6 @@ class System {
     std::queue<std::string> tx_queue;
 
 };
-int Flow2PWM(float);
+int Flow2PWM(double);
 
 #endif /* !MOTOR_HPP */
